requirements.txt
certifi                2024.6.2
charset-normalizer     3.3.2
easyocr                1.7.1
filelock               3.14.0
fsspec                 2024.6.0
idna                   3.7
imageio                2.34.1
intel-openmp           2021.4.0
Jinja2                 3.1.4
lazy_loader            0.4
MarkupSafe             2.1.5
mkl                    2021.4.0
mpmath                 1.3.0
networkx               3.3
ninja                  1.11.1.1
numpy                  1.26.4
opencv-contrib-python  4.10.0.82
opencv-python          4.10.0.82
opencv-python-headless 4.10.0.82
packaging              24.1
paho-mqtt              2.1.0
pillow                 10.3.0
pip                    24.0
pyclipper              1.3.0.post5
python-bidi            0.4.2
PyYAML                 6.0.1
requests               2.32.3
scikit-image           0.23.2
scipy                  1.13.1
setuptools             65.5.0
shapely                2.0.4
six                    1.16.0
sympy                  1.12.1
tbb                    2021.12.0
tifffile               2024.5.22
torch                  2.3.1
torchvision            0.18.1
typing_extensions      4.12.2
urllib3                2.2.1
